+++ lib/cli.js
@@ -1,7 +1,7 @@
 const inquirer = require('inquirer');
 const stringToFile = require('./stringToFile');
 const { validateHTMLColorName, validateHTMLColorHex } = require('validate-color');
-const Shape = require('./shapes');
+const { Circle, Square, Triangle } = require('./shapes');
 
 class CLI {
     constructor() {
@@ -85,12 +85,25 @@ class CLI {
             const file = answers.text + '.svg';  // TODO change this to be answers.text-logo.svg.
             
             // TODO switch to call correct shape class based on answers
-            // TODO remove shapeTest
-            const shapeTest = new Shape(answers);
 
-            return stringToFile(directory, file, shapeTest.renderSVG());
+            switch (answers.shape) {
+                case 'circle':
+                    const circle = new Circle(answers);
+                    return stringToFile(directory, file, circle.renderSVG());
+                    break;
+                case 'square':
+                    const square = new Square(answers);
+                    return stringToFile(directory, file, square.renderSVG());
+                    break;
+                case 'triangle':
+                    const triangle = new Triangle(answers);
+                    return stringToFile(directory, file, triangle.renderSVG());
+                    break;
+                default:
+                    break;
+            }
         } catch (err) {
-            console.err(err);
+            console.log(err);
         }
 
     }
