+++ lib/shapes.js
@@ -4,6 +4,7 @@ class Shape {
         this.textColor = answers.textColorName || answers.textColorHex;
         this.shapeColor =  answers.shapeColorName || answers.shapeColorHex;
     }
+
     renderShape() {
         const methodWarning = 'THIS METHOD MUST BE CALLED ON A CHILD CLASS WITH ITS OWN METHOD DECLARATION';
         console.log(methodWarning);
@@ -22,4 +23,34 @@ class Shape {
     }
 }
 
-module.exports = Shape;
+class Circle extends Shape {
+    constructor( answers ) {
+        super( answers );
+    }
+    
+    renderShape() {
+        return `<circle cx="150" cy="100" r="80" fill="${this.shapeColor}" />`
+    }
+}
+
+class Square extends Shape {
+    constructor( answers ) {
+        super( answers );
+    }
+    
+    renderShape() {
+        return `<rect x="90" y="40" width="120" height="120" fill="${this.shapeColor}" />`;
+    }
+}
+
+class Triangle extends Shape {
+    constructor( answers ) {
+        super( answers );
+    }
+    
+    renderShape() {
+        return `<polygon points="150, 18 244, 182 56, 182" fill="${this.shapeColor}" />`;
+    }
+}
+
+module.exports = Circle, Square, Triangle;
